move getAreas + getFocuses
